+++ scripts/macros.js
@@ -177,6 +177,7 @@ import {wrathOfTheStorm} from './macros/classFeatures/cleric/tempestDomain/wrath
 import {zealousPresence} from './macros/classFeatures/barbarian/zealot/zealousPresence.js';
 import {zombie} from './macros/monsterFeatures/zombie/zombie.js';
 import {zoneOfTruth} from './macros/spells/zoneOfTruth.js';
+import {danseMacabre} from './macros/spells/danseMacabre.js';
 export async function onHitMacro(workflow) {
 	if (workflow.targets.size === 0) return;
 	workflow.targets.forEach(async token => {
@@ -404,5 +405,6 @@ export let macros = {
 	'witherAndBloom': witherAndBloom,
 	'wrathOfTheStorm': wrathOfTheStorm,
 	'zealousPresence': zealousPresence,
-	'zoneOfTruth': zoneOfTruth
+	'zoneOfTruth': zoneOfTruth,
+	'danseMacabre': danseMacabre
 }